same filt coch as EEG + unsync erp start time fix
Bad practice but I'm commiting 2 independent changes.

One is adding bandpass filter with exactly the same parameters as the
EEG data and using it on the cochlear model. This is to make sure that
the filter alone cannot explain the difference between EEG and cochlear
model.

There's also a small inconsistency in the figure showing cycle-ERP for
the unsyncopated rhythm. The rhythmic pattern was x..xxx.xxx.. but we've
been consistently plotting this as xxx.xxx.xx.. because that's the most
common phase people tap. In order to be consistent across studies, we
can just plot the ERP starting from event 4 of the unsyncopated rhythm.